Adds lib/batches
