+++ package.json
@@ -24,7 +24,7 @@
   },
   "ava": {
     "files": [
-      "test/*.test.js"
+      "test/**/*.test.js"
     ],
     "verbose": true
   }
